placements/src/main/java/com/accolite/placements/models/ChangePassword.java
package com.accolite.placements.models;

public class ChangePassword {
	private String currentPassword;
	private String newPassword;
	
	public ChangePassword(String currentPassword, String newPassword) {
		super();
		this.currentPassword = currentPassword;
		this.newPassword = newPassword;
	}
	public ChangePassword() {
		super();
	}
	public String getCurrentPassword() {
		return currentPassword;
	}
	public void setCurrentPassword(String currentPassword) {
		this.currentPassword = currentPassword;
	}
	public String getNewPassword() {
		return newPassword;
	}
	public void setNewPassword(String newPassword) {
		this.newPassword = newPassword;
	}
}
